Add files via upload
fix: Correct the graphs and Optimal Policy question

After fixing the get_maxQ and choose_action functions,
the current graphs do not reflect the correct training of the
agent. A new training is required from scratch to reflect the
changes and the graphs should reflect the new training.

There is an error in the answer of Optimal Policy question. 
The error is in recognizing a sub-optimal policy as optimal policy.
It is corrected here.